+++ src/clojure_lsp/feature/completion.clj
@@ -220,7 +220,7 @@
                (mapv
                  (fn [element]
                    (let [require-edit (some-> cursor-loc
-                                              (r.transform/add-known-libspec (symbol (str (:alias element)))
+                                              (r.transform/add-known-alias (symbol (str (:alias element)))
                                                                              (symbol (str (:to element)))
                                                                              db)
                                               r.transform/result)]
@@ -238,7 +238,7 @@
              (map
                (fn [[element-ns completion-item]]
                  (let [require-edit (some-> cursor-loc
-                                            (r.transform/add-known-libspec (symbol cursor-alias) element-ns db)
+                                            (r.transform/add-known-alias (symbol cursor-alias) element-ns db)
                                             r.transform/result)]
                    (cond-> completion-item
                      (seq require-edit) (assoc :additional-text-edits (mapv #(update % :range shared/->range) require-edit)))))))))))