+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Editor
   - Fix "Add require" code actions adding multiple requires instead of the selected.
   - Improve "Add require" wording, making it easier to understand what each different action will do.
+  - Smart check all available refers to require, adding refer options to `Add require` code actions. #627
 
 This release was supported by [Clojurists Together](https://www.clojuriststogether.org/)